Began fixing a bug breaking self-class identification
The class toolkit does not correctly display the class of the bot that's
running the check. This means that a bot can know the class of any
allied bot except for itself (I know this is completely replaceable by
normal blocks, but I'm doing this for the sake of completeness). This
commit is the first step towards fixing this bug, will continue work
tommorow