Bug fix: Greatly improved broadcast consistency and fixed missed messages
To understand this bug, I'll need to describe how plugins are handled:

Every 10 ticks (~1 second) a botReport broadcast is made. Whenever one is made, bots will ping back their status and run each plugin. The problem with this is that plugins also handle the reciever logic, which must always be running in order to consistently recieve messages. Likewise, the broadcasters inside of the plugins were set to broadcast every xth tick. But the problem is that the plugin already only runs every 10th tick, so a broadcaster set to broadcast every 3rd tick would never broadcast at all (since the 10th ticks are never a multiple of 3). This also lead to broadcasts activating at seemingly random times. This fix restructures plugins so that they are left to delegate their own reciever and broadcast functions.